@@ -41,24 +41,22 @@ class Collection extends React.Component<Props> {
     numberOfColumns: 4,
     refreshing: false,
   };
-  constructor(props: Props) {
-    super(props);
-    this.styles = getStyles(props);
-  }
-
-  styles: Object;
+  itemStyle = {
+    height: width / (this.props.numberOfColumns || 4),
+    width: width / (this.props.numberOfColumns || 4),
+  };
 
   renderCell = ({ item }: { item: * }) => {
     const child = this.props.renderItem(item);
     if (!child) return null;
     return React.cloneElement(child, {
-      style: StyleSheet.flatten([child.props.style, this.styles.item]),
+      style: StyleSheet.flatten([child.props.style, this.itemStyle]),
     });
   };
 
   renderRow = ({ item }: { item: * }) => (
     <FlatList
-      style={this.styles.wrapper}
+      style={styles.wrapper}
       numColumns={this.props.numberOfColumns}
       renderItem={this.renderCell}
       data={item}
@@ -67,7 +65,6 @@ class Collection extends React.Component<Props> {
   );
 
   genListSection = (data: Data) =>
-    // $FlowFixMe
     data.map(item => ({ ...item, data: [item.data] }));
 
   render() {
@@ -115,16 +112,9 @@ class Collection extends React.Component<Props> {
 
 export default withTheme(Collection);
 
-const getStyles = (props: Props): Object => {
-  const { numberOfColumns = 4 } = props;
-  return StyleSheet.create({
-    wrapper: {
-      flexDirection: 'row',
-      flexWrap: 'wrap',
-    },
-    item: {
-      height: width / numberOfColumns,
-      width: width / numberOfColumns,
-    },
-  });
-};
+const styles = StyleSheet.create({
+  wrapper: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+  },
+});
