@@ -5,12 +5,8 @@ import type { Theme } from '../types/Theme';
 
 const DarkTheme: Theme = {
   primaryColor: colors.orange,
-  primaryLightColor: color(colors.orange)
-    .lighten(0.5)
-    .string(),
-  disabledColor: color(colors.orange)
-    .lighten(0.8)
-    .string(),
+  primaryLightColor: color(colors.orange).lighten(0.5).string(),
+  disabledColor: color(colors.orange).lighten(0.8).string(),
   backgroundColor: 'rgb(17,29,33)',
   barColor: 'rgb(29,58,63)',
   dividerColor: colors.greyD2,
