@@ -8,12 +8,8 @@ import type { Theme } from '../types/Theme';
  */
 const DefaultTheme: Theme = {
   primaryColor: colors.blue,
-  primaryLightColor: color(colors.blue)
-    .lighten(0.5)
-    .string(),
-  disabledColor: color(colors.blue)
-    .lighten(0.8)
-    .string(),
+  primaryLightColor: color(colors.blue).lighten(0.5).string(),
+  disabledColor: color(colors.blue).lighten(0.8).string(),
   backgroundColor: colors.white,
   barColor: colors.greyL2,
   dividerColor: colors.grey,
